refactor: Turn LLMQType into `enum class`

+++ src/rpc/server.cpp
@@ -266,7 +266,7 @@ void CRPCTable::InitPlatformRestrictions()
         {"getblockhash", {}},
         {"getblockcount", {}},
         {"getbestchainlock", {}},
-        {"quorum", {"sign", Params().GetConsensus().llmqTypePlatform}},
+        {"quorum", {"sign", (uint8_t)Params().GetConsensus().llmqTypePlatform}},
         {"quorum", {"verify"}},
         {"verifyislock", {}},
     };
